requirements/base.txt
django==5.1.7
psycopg2==2.9.3
black==25.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
drf-spectacular==0.28.0